Updated Verasonics examples to use single for channel_data.
Former-commit-id: dc0c6839a4055f0bdb08a68f99ad192a689c25da

+++ @verasonics/verasonics.m
@@ -137,9 +137,9 @@ classdef verasonics < handle
 
         
         function data_out = time_shift_data(h,data_in,t_in,t_out,interpolation_factor,channel_data)
-            % First do a sinc interpolation
+            % First do a interpolation to avoid bug described in Issue #16
             t_in_interp = linspace(t_in(1),t_in(end)+((interpolation_factor-1)/interpolation_factor)*(1/channel_data.sampling_frequency),length(t_in)*interpolation_factor); 
-            data_tx_interpolated = interpft(double(data_in),length(t_in)*interpolation_factor);
+            data_tx_interpolated = single(interpft(double(data_in),length(t_in)*interpolation_factor));
             %%
             %                     channel = 64;
             %                     figure(99);clf;hold all;
@@ -149,8 +149,8 @@ classdef verasonics < handle
             %                     subplot(212);hold all
             
             %%
-            % read data
-            data_out=interp1(t_in_interp,data_tx_interpolated,t_out,'linear',0);
+            % then do the 
+            data_out=single(interp1(t_in_interp,data_tx_interpolated,t_out,'linear',0));
         end
     end
 end
